Issue/616: refactor: Extracted sending to chat functionality to new utility class
