+++ presentation/component/button-send-to-chat/button-send-to-chat-viewmodel.mjs
@@ -1,10 +1,6 @@
 import ButtonViewModel from "../button/button-viewmodel.mjs";
-import { VISIBILITY_MODES } from "../../chat/visibility-modes.mjs";
-import DynamicInputDialog from "../../dialog/dynamic-input-dialog/dynamic-input-dialog.mjs";
-import DynamicInputDefinition from "../../dialog/dynamic-input-dialog/dynamic-input-definition.mjs";
-import { DYNAMIC_INPUT_TYPES } from "../../dialog/dynamic-input-dialog/dynamic-input-types.mjs";
-import { ChatUtil } from "../../chat/chat-utility.mjs";
 import { ValidationUtil } from "../../../business/util/validation-utility.mjs";
+import SendToChatHandler from "../../utility/send-to-chat-handler.mjs";
 
 /**
  * A button that allows sending a document or one of its properties to the chat. 
@@ -96,61 +92,16 @@ export default class ButtonSendToChatViewModel extends ButtonViewModel {
   }
 
   /**
-   * @param {Event} event
-   * 
    * @async
-   * @protected
    * @override
+   * @protected
    */
-  async _onClick(event) {
+  async _onClick() {
     if (this.isEditable !== true) return;
 
-    const nameInputVisibility = "nameInputVisibility";
-
-    const dialog = await new DynamicInputDialog({
-      localizedTitle: game.i18n.localize("system.roll.query"),
-      inputDefinitions: [
-        new DynamicInputDefinition({
-          type: DYNAMIC_INPUT_TYPES.DROP_DOWN,
-          name: nameInputVisibility,
-          localizedLabel: game.i18n.localize("system.general.messageVisibility.label"),
-          required: true,
-          defaultValue: VISIBILITY_MODES.asChoices().find(it => it.value === VISIBILITY_MODES.public.name),
-          specificArgs: {
-            options: VISIBILITY_MODES.asChoices(),
-          }
-        }),
-      ],
-    }).renderAndAwait(true);
-    
-    if (dialog.confirmed !== true) return;
-
-    const visibilityMode = VISIBILITY_MODES.asArray().find(it => it.name === dialog[nameInputVisibility].value);
-    
-    if (this.propertyPath !== undefined) {
-      if (this.target.sendPropertyToChat !== undefined) {
-        this.target.sendPropertyToChat(this.propertyPath, visibilityMode);
-      } else {
-        ChatUtil.sendPropertyToChat({
-          obj: this.target,
-          propertyPath: this.propertyPath,
-          parent: this.target,
-          actor: this.actor,
-          visibilityMode: visibilityMode
-        });
-      }
-    } else {
-      if (this.target.sendToChat !== undefined) {
-        this.target.sendToChat(visibilityMode);
-      } else {
-        ChatUtil.sendPropertyToChat({
-          obj: this.target,
-          propertyPath: this.propertyPath,
-          parent: this.target,
-          actor: this.actor,
-          visibilityMode: visibilityMode
-        });
-      }
-    }
+    await new SendToChatHandler().prompt({
+      target: this.target,
+      dialogTitle: this.chatTitle,
+    });
   }
 }
